chore: Upgrade actions/upload-artifact to v4 in build workflow
This commit updates the `actions/upload-artifact` action from version 3 to version 4 in the `.github/workflows/build-and-release.yml` file. This upgrade applies to both the Debian package and Docker files artifact uploads.

The purpose of this change is to utilize the latest features, bug fixes, and security improvements offered by version 4 of the action. Keeping the workflow up-to-date with the latest action versions ensures better performance and reliability.