feat: Add 128x128 icon for improved app manifest support
This commit adds a 128x128 pixel icon (icon-128.png) to the 4chat/assets directory. This icon is intended to be used within the application manifest (e.g., for PWA support) to provide a higher resolution icon for various platforms and devices.  The inclusion of this icon improves the visual appearance and user experience when the application is installed or added to the home screen.