feat: Add 64x64 application icon for improved UX
This commit adds a 64x64 pixel PNG icon for the 4chat application.
This icon will be used as a higher-resolution favicon and app icon,
enhancing the user experience by providing a clearer and more visually
appealing representation of the application across different platforms
and devices. The new `icon-64.png` file is located in the `4chat/assets`
directory.